Extract SchedulePageClient; make page server
Move the client-side schedule UI into a new components/lecturer/SchedulePageClient.tsx file.
 Convert app/lecturer/schedule/page.tsx into a server component
